Imported SP changes and vpc's. Modified VPC to remove regkey requirement

+++ src/game/client/hl2/hud_damageindicator.cpp
@@ -176,10 +176,10 @@ void CHudDamageIndicator::DrawDamageIndicator(int side)
 
 	int insetY = (m_flDmgTall1 - m_flDmgTall2) / 2;
 
-	int x1 = m_flDmgX;
-	int x2 = m_flDmgX + m_flDmgWide;
-	int y[4] = { (int)m_flDmgY, (int)(m_flDmgY + insetY), (int)(m_flDmgY + m_flDmgTall1 - insetY), (int)(m_flDmgY + m_flDmgTall1) };
-	int alpha[4] = { 0, 1, 1, 0 };
+	float x1 = m_flDmgX;
+	float x2 = m_flDmgX + m_flDmgWide;
+	float y[4] = { m_flDmgY, m_flDmgY + insetY, m_flDmgY + m_flDmgTall1 - insetY, m_flDmgY + m_flDmgTall1 };
+	float alpha[4] = { 0.0f, 1.0f, 1.0f, 0.0f };
 
 	// see if we're high damage
 	bool bHighDamage = false;
